Move auth to root
